requirements.txt
charset_normalizer
certifi
idna
requests
pyhcl
urllib3